refactor: separate header and footer style files
